+++ witdim_frontend-web/CHANGELOG.md
@@ -7,6 +7,14 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2021-09-28
+
+- Handle empty `item[]` on `/home` route
+- `DELETE` and `CREATE` endpoints are functional
+- Heavy work on api endpoints
+- Work on `/add` route
+- UI still heavily in development
+
 ## [0.0.5] - 2021-09-28
 
 - Added test svg buttons on utilbar